requirements.txt
aiofiles==23.2.1
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
auto-py-to-exe==2.34.0
autopep8==2.0.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.12.25
bottle-websocket==0.2.9
build==1.0.3
CacheControl==0.13.1
cachetools==5.5.1
capstone==6.0.0a3
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.8
colorama==0.4.6
colored-traceback==0.4.2
comtypes==1.1.14
crashtest==0.4.1
cryptography==44.0.2
cssselect==1.2.0
distlib==0.3.8
distro==1.9.0
Django==5.1.4
django-ninja==1.3.0
dulwich==0.21.7
Eel==0.16.0
fastapi==0.115.8
fastjsonschema==2.19.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.13.1
Flask==2.0.1
flask-cors==3.0.10
fsspec==2025.2.0
future==0.18.3
#gevent==22.10.2
#gevent-websocket==0.10.1
google-auth==2.38.0
google-genai==1.2.0
groq==0.18.0
goose3==3.1.19
gradio==5.16.1
gradio_client==1.7.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
intervaltree==3.1.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
keyring==24.3.0
langdetect==1.0.9
lxml==5.3.1
lxml_html_clean==0.4.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.3
olefile==0.46
openai==1.63.2
orjson==3.10.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
paramiko==3.5.1
pdfminer.six==20250327
pefile==2023.2.7
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.9.6
platformdirs==3.11.0
plumbum==1.9.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pwntools==4.14.1
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautocad==0.2.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.22.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyelftools==0.32
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rapidfuzz==3.5.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
retry==0.9.2
ROPGadget==7.6
rpyc==6.0.1
rsa==4.9
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
selenium==4.28.1
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.2
starlette==0.45.3
sympy==1.13.1
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
tomlkit==0.12.3
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.1
trove-classifiers==2023.11.29
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
unicorn==2.1.4
unix-ar==0.2.1
urllib3==2.1.0
uvicorn==0.34.0
virtualenv==20.25.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.0.3
whichcraft==0.6.1
wsproto==1.2.0
xgboost==2.1.4
yara-python==4.5.1
zipp==3.17.0
zope.event==4.6
zope.interface==6.0
zstandard==0.23.0
pypdf2==1.26.0
oletools==0.60
reportlab==4.2.5